ls コマンドを作る5 を暫定実装した
